d_message OK (via dFlow_c)

+++ include/d/d_sc_game.h
@@ -127,7 +127,7 @@ public:
     virtual int draw() override;
     virtual void deleteReady() override;
 
-    void triggerExit(s32 room, u8 exitIndex, s32 forcedNight = 2, s32 forcedTrial = 2);
+    void triggerExit(s32 room, u16 exitIndex, s32 forcedNight = 2, s32 forcedTrial = 2);
     // void triggerEntrance(const char *stageName, u8 room, );
     const LinkReloadInfo &getLinkReloadInfo() const;
